Disable Audio in Main Camera
Warnings abounded about having 2 audio listeners when there should only
every be one.  Disabling this in the standbyCamera on NetworkManager
removes this problem.  May be better to create a function later on to
manage doing this so we don't forget to do both and end up with a camera
that is on without audio.