Player Control over networking
Tutorial 5 added in controlling the character and sending updates over
the network.  Major changes in Unity 5 and PUN have led to some changes
needed in how things are implemented.

Capsule object added to FPSController so we can see the character in the
scene.

Also added room name to the GUI Text Label to verify we are in the same
room.

Removed extra Spawn Point from TeamB and also put the Spawn Points in
front of the buildings for both teams rather than behind for one team.